+++ changelog.md
@@ -1,5 +1,6 @@
 - Custom Ship AI can be enabled for ships in player fleet via a hullmod.
 - Added LunaLib setting to enable/disable needler AI attacking enemy ships only when their shields are raised.
+- Added integration with Leading PIP. When AI Tweaks Aim Assist is enabled, the targeting lead indicator centers on the selected target.
 - Added an option to remove the square grid on the combat map. This option is available in the LunaLib config under the Debug tab.
 
 Version 1.10.12